requirements.txt
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
openai>=1.12.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tqdm>=4.66.0 